Relax the domain on put-away rule destination
Currently, we may need to have a source going to the same destination.
Example: WER → WER

It's a bit weird but the reason is this: if we want to apply a chaotic
strategy when we move to WER, as the strategy is applied on a put-away
rule's destination, we need such put-away rule.

Later, we may want to decorrelate the standard put-away rules and the
storage type strategy: apply the put-away rule first to the destination,
then apply the storage type strategy of this location.