Rename 'device' to 'profile'
We don't want to force having one device record per hardware device,
since adding a new scanner would require to create a new device on Odoo.
Instead, we prefer to have profiles, which has to be selected at loading
of the client application. The profile will hold the configuration for
the interactions (warehouse, but maybe later printer, ...).

The allowed profiles can be restricted to groups. A user can be forced
to use a profile.