Add Actions Components to share business logic
Service Components are methods exposed to the REST API.
They need to share common methods for their business logic,
so enter the Actions Components to hold them.
Alternatively, we could have put the methods directly on the Models,
but at least here we don't crowd the models with many methods.

To create a new Action Component, the bare minimum is:

  from odoo.addons.component.core import Component

  class StockPackageLevelAction(Component):
      _name = "shopfloor.stock.package_level.action"
      _inherit = "shopfloor.process.action"
      _apply_on = "stock.package_level"

      def hello(self):
          return "world"

And to use it from any Service or Action component:

  self.actions_for("stock.package_level").hello()