Use a common Action Component for pack transfer 'validate'
The Action Copmonent contains all the support methods, used in the
workflow for the "validate" method in the Service Component.

The validate method is currently the same in the Service Component for
pack transfer and put-away, but it can make sense if we consider it
deals mainly with the responses to do for the client application: if we
want to, for instance, add a step for one process but not the other, we
want to change this is the Service Component, reusing the methods of the
Actions Component.