Rework how responses include data needed for the next state
The 'validator.response' components have to define a `_states()` method
that gives the match between a state and the schema a state requires.

Validators for methods which can make the client transition must declare
the list of possible next states using `next_states=["state_a",
"state_b"]` (which must exists in the `_states()` method).

When an endpoint returns data for a state, the data will now be enclosed
in a key with the same name as the state, this is in order to support
polymorphism in schemas (an endpoint being able to return different data
depending on the next state).

General idea of a schema for a method that changes state (data may vary,
in this example, next_state will be one of "confirm_start", "start",
"scan_location"):

{
  message {
    message_type* string
    message*      string
  }
  next_state      string
  data {
    confirm_start {...}
    start {...}
    scan_location {...}
  }
}

General idea of a schema for a generic method (data may vary):

{
  message {
    message_type* string
    message*      string
  }
  data {
    size*         integer
    records*      integer
  }
}