Add link between info screen without routing
The info screen can be accessed by two different ways. Using the
magnifying glass icon in the top bar (/scananything) or by
clicking a qestion mark button in the course of a scenario.

If through the scananything route the routing can be used it is not the
case through the scenario. This is why the handling of navigation
between multiple info screen is handled differentely.