pack putaway: fix test after change of behavior in odoo
After this commit, the behavior has changed:
https://github.com/odoo/odoo/commit/90ae4d6afe7

When we change the destination location of a package level, it
automatically updates the move to have the same destination location.

Before this commit, the dest. location of the package level was not
updated, now it is updated with the move line's dest, so the move is
updated in turn.