rest validators: add an implicit start state
Since all the endpoints could theoretically bring back to the start
state in case of unrecoverable error, we should not need to add the
state everywhere. It's still fine and probably better to keep the
start state in the possible next state of a state when it's the logical
next state (not because of an error) for documentation.