checkout: use common methods for getting data for endpoints
The idea is to share data and validation schemas for data returned
by endpoints, per model, across the different scenarii.

Currently, only the checkout uses it, but the new ones we implement
should use the same schemas and data methods as much as possible.
Ideally, we should rework the existing ones to use them.