Keep final destination on push routing rules
In case of a put-away, when using a push routing rule, we want
to keep the expected destination.

For instance, we have a move:

* Input -> Highbay (draft)

It's assigned and the put-away rules compute a move line going to
Highbay/X1Y2.

When a routing is applied, before this commit, it would add a move at
the end, it would look like:

* Input -> Handover (assigned)
* Handover -> Highbay (confirmed)

On reservation of the second move (Handover -> Highbay), it would
compute again the put away rules to find a place in the whole highbay.

With the change now, the moves would look like:

* Input -> Handover (assigned)
* Handover -> Highbay/X1Y2 (confirmed)

So the move line cannot go elsewhere.

To implement this, I decided to remove the 'routing_rule_id' field
stored on 'stock.move', as this field is only needed by the methods
_apply_routing_rule_pull/push and never afterwards. The original
location has to be propagated to these methods as well, so now there
is a single dict of values used to apply the rules.