Apply pre-commit
Check waiting moves in "Not Empty" put-away constraints

When an unreserved move's destination is already a bin, we assume:

* the move is probably waiting on another that will be processed soon
* as the destination of the move is already a bin (not Stock, ...), the
  good *will* arrive in this precise location and not another

So we prevent new move lines to target this location.

Note: as we do not know the lot yet in the move, we cannot check if we
have a planned lot for the same lot in the "Do not mix same lot": ignore
it.