backend: support multiple picking types in services
Except for the cases where we can create a move, a single menu should
be able to work on several picking types.

Later, some menus will have a flag "create_move", which add an extra
step in the scenario. In this case, we'll still have a selection of
several picking types on the menu, but an additional M2o to select the
picking type used for creation of moves.