cluster picking: reassign moves after stock issue
In the process, I had to change the sorting of the lines to reflect
the sorting of the models, which broke the test checking that on
the split of a move line because we picked not enough qty, the next
line to pick should be the split line. It was working by chance,
so I added a parameter to force pick the line.