Add option to disable zero check in scenario
Currently, it applies only for Cluster Picking, it will apply
on Zone Picking and Discrete Order Picking too.

When a location becomes empty and the option is active, the user has
to validate that the location is really empty. It doesn't really make
sense for picking types that move pallets, so they would disable it
there.

There is an ongoing discussion whether this option should be on the
picking type or on the shopfloor menu, it is possible that this option
will be moved.