Raise priority only if the direct next move has consolidation
In the following graph:

                  PICK/001 ━►  PACK/001  ┓
                                         ┃
                  PICK/002 ┓             ┣━► OUT/001
                           ┣━► PACK/002  ┛
       INT/001 ━► PICK/003 ┛

If we have the consolidate flag on "PACK", we want to raise the priority
only when we start to move goods in PACK/001 or PACK/002 (not INT/001),
because this is in the packing zone that we are limited in space.

Also, when for instance a move of PICK/002 or PICK/003 is set to done,
*any* move (all products) that go to PACK/002 are concerned by the
priority raise: we want to finish the transfer.