Make stock_storage_type depend on stock_putaway_hook
The putaway based on storage type *must* be applied at the very end
of the putaway rules, because it depends on the putaway location found.

If we execute it by calling super() and we added different putaway
strategies in StockLocation, we have no way to be sure it is the last
called.

That's why now it depends on stock_putaway_hook, which allow to search
for alternative put-away hooks, and call _putaway_strategy_finalizer at
the very end, where we can plug the storage type rules.