Add check of scanned dest. location vs move's destination
When we scan a location, generally, we have these rules:

* If the scanned location is not a child (or =) of the picking type's
  default destination, reject it
* If the scanned location is not a child (or =) of of the current move
  line's destination, ask for a confirmation

However, we have no check on the move's destination location.

Example
-------

Locations:

Packing Zone
 \- PACK-1
 \- PACK-2

Our transfer has:

* Picking type destination location: Packing Zone
* Move destination location: Packing Zone/PACK-1
* Move line destination location: Packing Zone/PACK-1

Current behavior:

* If we scan "PACK-1", the location is accepted directly
* If we scan "PACK-2", the app asks for a confirmation

If the user confirms PACK-2, we'll have inconsistencies with such
result:

* Move destination location: Packing Zone/PACK-1
* Move line destination location: Packing Zone/PACK-2

Also, the next move in the chain expects to receive goods in "Packing
Zone/PACK-1", so we break the chain.

Expected behavior after this commit:

* If we scan "PACK-1", the location is accepted directly
* If we scan "PACK-2", the location is rejected

Considerations
--------------

This issue was detected in the location content transfer, but it may
happen in the other scenarios, so all of them should be fixed.

Except for "Single Pack Transfer" which was already handling this in a
different way, by updating the move's destination location. I don't
think it is correct, as I doubt the next move in the chain will be
correct then. But I don't want to break a use case that could be used,
so not changing it for now.

In Zone Picking, I simplified a bit the checks on the location: by
checking *first* if we are in the picking type (and in the move
desination now), we can return early and avoid repeating the condition
on the next check.