front: save space for main doc title on small screens
When the main doc title is in the app header (eg: checkout picking)
the scan anything icon is hidden for screens below 320px width.

Font size for doc title is decreased by 0.05rem which gives even more
space for extra digits (currently 22 max).

The action is now always available in the main menu.