checkout: Fix flaky test for list_packaging
CheckoutListSetPackagingCase.test_list_packaging_ok is randomly failing,
because it relies on the order of the records which indeterminate.

As shopfloor depends on the module "product_packaging_type", which
changes the default ordering of "product.packaging" to:

    _order = "product_id, type_sequence"

The "sequence" on the packaging is useless.
Create packaging types so the order is stable.