new method to validate moves in a separate transfer (#72)
Calling _action_done() directly on the stock.move while leaving the
stock.picking "assigned" with the other moves already required various
workarounds as it is normally not an expected flow in Odoo.

This method is to be used in services instead of calling _action_done() on
the moves. If a move to set to done contains move lines that should not
be validated, the line to keep should be extracted beforehand with
`move_id.split_other_move_lines`.