backend: remove code related to _action_done usage
Calling _action_done() directly on the stock.move while leaving the
stock.picking "assigned" with the other moves already required various
workarounds as it is normally not an expected flow in Odoo.

We have detected new issues with package levels which alter "done" move
lines when we call _action_assign() on the other moves again.

Previous commits removed the calls to _action_done() on the moves and
replaced them by a method that extract the move to set to done in a new
stock.picking (StockMove.extract_and_action_done()).

This commit removes the hacks which are no longer used.