backend: fix reset of result_package_id on move._action_assign()
This is actually an issue in the core stock module, which is fixed
locally as it is much more likely to happen in shopfloor than any
other places (and the consequences if it happens in shopfloor have
no workaround, while you can still edit the move lines when it happens
on the UI).

Reported as https://github.com/odoo/odoo/issues/58491
The details about the issue are there.

In short: calling _action_assign on a stock.move create/write package
levels which may in turn replace the "result_package_id" of a move line,
even if it was already set by a user.

To prevent this, we skip the package level sync when:

* any line related to the package has a result_package_id set
* any line related to the package has a qty_done set

Depending on the response of odoo to this issue, we may later remove
this commit partially or completely.