location transfer: implement unreserve of existing moves
When we scan a location which already have reserved move lines for other
picking types, and the option is active on the menu, unreserve the moves,
create new moves for the location content transfer and re-assign the
unreserved moves.

If one of the move that would be unreserved has already been picked
(qty_done > 0), an error is returned.

I added a savepoint to rollback when the we had to unreserve moves and
finally we can't reserve the location transfer moves for any reason. As
a side effect, we don't need anymore to cancel them as they will be
rollbacked.