single pack transfer: implement unreserve of existing moves
When we scan a package which already have reserved move lines for other
picking types, and the option is active on the menu, unreserve the moves,
create new moves for the location content transfer and re-assign the
unreserved moves.

If one of the move that would be unreserved has already been picked
(qty_done > 0), an error is returned.