location transfer: implement ignore no putaway available option
When we start a location content transfer, the moves (created or not by
the menu) are assigned. When we use this menu with putaways, we can
activate the option "ignore no putaway available".
When no putaway location is available, the destination location
remains the default one. If this option is active and the destination
location didn't change: rollback the action_assign. This will prevent
having move lines for things we don't know where to place.