Add stock_dynamic_routing_reserve_rule
Glue module between stock_dynamic_routing and stock_reserve_rule.
It adds a link to open reserve rules from dynamic routings, and
displays warning on dynamic routings when there are risks of
inconsistencies because of discrepancies between dynamic routing
and reservation rules.

For the record, these issues are:

How the application of the dynamic routing works:

1. StockMove._action_assign() is executed in a savepoint, to see what move
   lines are created and what are their source and destinations
2. If no dynamic routing matches the picking type and one of the lines:
   source / location (+ domain), the savepoint is released (we keep the
   initial _action_assign)
3. If we have a dynamic routing, the savepoint is rollbacked, we apply
   the dynamic routing (split moves if required, change picking type, add
   chained move), and call again StockMove._action_assign()

As in step 3, the picking type can be changed, and the reservation rules
are dependent of the picking type, if we have a reservation rule for the
initial picking type and no reservation rule for the new one, we may
get a different reservation which does no longer takes the good at the
place expected initially.