Store release channel on stock.move
When using stock_dynamic_routing, the moves are reassigned to a new
picking (with StockMove._assign_picking()), and the original picking is
cancelled if it contains no moves anymore. As we want to keep

Note: maybe we should extend the assign picking mechanism to never
assign a move in a picking with a different release channel (channel is
empty or identical). As we extend it in
stock_picking_group_by_partner_by_carrier, it should be a new
dependency/glue module on it, or better, only extract the hook added
by stock_picking_group_by_partner_by_carrier for the domain in a base
module, used by both.