[FIX] stock_reception_screen: create package on step validation
Before this commit the package was created on the fly thanks to the
'inverse' method on the `current_move_line_package` field.
This was generating some issues regarding sync of data + the fact that
unused packages were created if the user:

1. fills up the package name and hit the "Cancel" or "Exit" button (which are
   actually saving the form, and so triggers the inverse method creating the
   package),

2. get back on the screen and enter/scan another package name,
   overwriting the existing one (which is now like a "ghost" package not
   related anymore to the move line).

Now if the user fills up the package name and save the form ("Cancel" or "Exit")
we simply store the name of the package, and we wait for the user to
validate the current step through the "Next" or "Next package" buttons,
which will create the package.