Ensure moves routed from different location are reserved before other
Take this use case:

* When we pick goods in HB, we have to go through Handover
* When we pick goods and their priority is 3, we change the picking type
  to "PRIO"

We have a Pick transfer for 30 units. We have 12 units in Stock/HB, 18 in
Stock/Shelving, priority is 3.

When we reserve, the dynamic routing's code does:

* run _action_assign on the move in a savepoint, to peek where we can
  reserve, and rollback the savepoint
* according to the reservations, split the move in 12 and 18
* apply the routings for the units:

  * in shelving, the move picking type is changed to PRIO
  * the other move is updated to take goods from HB to Handover in the
    Handover picking type, and a new waiting move is created for Handover
    -> Output with the PRIO picking type (will be in the same transfer as
    the shelving move)

If case the reservation of the shelving move is done first, it could
reserve the goods in HB that were reserved for the other move. We end
up with: HB move without reservation, the shelving stock taking goods in
HB without going through handover.

To prevent moves of a less specific location (such as Stock) stealing
moves of a more specific location (such as Stock/HB), sort the moves
to assign by reserved parent path of their source locations, as we know
the most specific locations cannot reserve moves above them. The order
within one location is preserved.