[IMP] shopfloor: create normal backorder when validating moves
Before: when moves are processed, we were managing two cases:

- if the current moves to process are all the moves remaining in a
  transfer, we validate the transfer

- if the current moves to process are a subset of available moves of a
  transfer, then we put these moves to process in a new transfer which
  is validated. All remaining moves stay in the current transfer.

After: a third case is added:

- if the current moves to process are equal to all available moves of
  the current picking - but there are still unavailable moves to process -
  then we validate the current transfer as usual, which will create a
  backorder with remaining/unavailable moves.

As this third case is very specific to Shopfloor, it has been
implemented in a new StockAction component (used by all related scenarios).
Now all scenarios use this new component to validate stock moves,
and the `extract_and_action_done` method on `stock.move` can stay generic.
The opportunity was taken to create a new method `split_unavailable_qty`
on `stock.move` to put unavailable qty to a new move and update the
current move to `assigned` (it was done directly in
`extract_and_action_done` before).