Optimize "will contain" computed fields on stock.location
The 2 fields:

* location_will_contain_product_ids
* location_will_contain_lot_ids

Are used only when applying a related location storage type with
"do_not_mix_products" or "do_not_mix_lots". As the result of the
computed fields are stored in m2m related fields, we can prevent a
lot of insertions in these relation tables by not computing them
if they have no location storage type that will need them.